Completely overhauled scanning page and workflow (does not work yet...)
