+++ app/(tabs)/_layout.tsx
@@ -1,4 +1,4 @@
-import { Foundation } from "@expo/vector-icons";
+import { Ionicons } from "@expo/vector-icons";
 import { Tabs } from "expo-router";
 import React from "react";
 
@@ -11,7 +11,7 @@ export default function TabLayout() {
           headerShown: false,
           title: "Home",
           tabBarIcon(props) {
-            return <Foundation name="home" size={24} color={props.color} />;
+            return <Ionicons name="home" size={24} color={props.color} />;
           },
         }}
       />
@@ -22,13 +22,7 @@ export default function TabLayout() {
           headerTitle: "Scan Receipt",
           title: "Scanner",
           tabBarIcon(props) {
-            return (
-              <Foundation
-                name="magnifying-glass"
-                size={24}
-                color={props.color}
-              />
-            );
+            return <Ionicons name="scan" size={24} color={props.color} />;
           },
         }}
       />
