+++ app/(tabs)/scan-receipt/index.tsx
@@ -1,7 +1,8 @@
 import { Ionicons } from "@expo/vector-icons";
 import { AutoFocus, Camera, CameraType } from "expo-camera";
 import * as ImageManipulator from "expo-image-manipulator";
-import { Link, Stack } from "expo-router";
+import * as ImagePicker from "expo-image-picker";
+import { router } from "expo-router";
 import React, { useState, useEffect } from "react";
 import {
   Text,
@@ -16,12 +17,10 @@ import {
 
 export default function Scanner() {
   const [hasPermission, setHasPermission] = useState<boolean | null>(null);
-  const [capturedImage, setCapturedImage] = useState<string | null>(null);
-  const [detectedText, setDetectedText] = useState<string>("");
-  const [output, setOutput] = useState<string>("");
+  const [selectedImage, setSelectedImage] = useState<string | null>(null);
   const [cameraType, setCameraType] = useState<CameraType>(CameraType.back);
 
-  const { width, height } = Dimensions.get("window");
+  let camera: Camera | null = null;
 
   useEffect(() => {
     (async () => {
@@ -30,6 +29,7 @@ export default function Scanner() {
     })();
   }, []);
 
+  // Take a picture using Expo Image Manipulator
   const takePicture = async () => {
     if (!camera) return;
 
@@ -38,83 +38,7 @@ export default function Scanner() {
       compress: 0.5,
     });
 
-    setCapturedImage(photo.uri);
-
-    const apiKey = "AIzaSyDFOfZ6SLPXEoDvF7RqdML5NXxOfySeKa4"; // Replace with your API key
-    const endpoint = `https://vision.googleapis.com/v1/images:annotate?key=${apiKey}`;
-    const base64Image = await convertImageToBase64(photo.uri);
-
-    fetch(endpoint, {
-      method: "POST",
-      body: JSON.stringify({
-        requests: [
-          {
-            image: {
-              content: base64Image,
-            },
-            features: [
-              {
-                type: "TEXT_DETECTION",
-              },
-            ],
-          },
-        ],
-      }),
-    })
-      .then((response) => response.json())
-      .then((data) => {
-        const annotations = data.responses[0].textAnnotations;
-        if (annotations && annotations.length > 0) {
-          setDetectedText(annotations[0].description);
-          setOutput(parseHEBReceipt(annotations[0].description)); // Parse the detected text
-        } else {
-          setDetectedText("No text detected");
-          setOutput("");
-        }
-      })
-      .catch((error) => console.error("Error:", error));
-  };
-
-  const convertImageToBase64 = async (imageUri: string) => {
-    const response = await fetch(imageUri);
-    const blob = await response.blob();
-    const base64String = await blobToBase64(blob);
-    return base64String;
-  };
-
-  const blobToBase64 = async (blob: Blob) => {
-    return new Promise<string>((resolve, reject) => {
-      const reader = new FileReader();
-      reader.onerror = reject;
-      reader.onload = () => {
-        const result: string | null = reader.result as string;
-        if (result) {
-          const [, base64Data] = result.split(",");
-          resolve(base64Data);
-        } else {
-          reject(new Error("Failed to read file as base64."));
-        }
-      };
-
-      reader.readAsDataURL(blob);
-    });
-  };
-
-  let camera: Camera | null = null;
-
-  const parseHEBReceipt = (text: string) => {
-    // RegExp to match 1-2 digits followed by whitespace followed by any characters except "Ea."
-    const regex = /(^\d{1,2})\s+((?!Ea\.).*)/gm;
-    const matches = [...text.matchAll(regex)];
-
-    // Maps the matches to an array to print name
-    const items = matches
-      .map((item, index) => {
-        const [, , name] = item;
-        return `${index + 1}) ${name}`;
-      })
-      .join("\n");
-    return items;
+    setSelectedImage(photo.uri);
   };
 
   // Auto Focus code borrowed from: https://github.com/expo/expo/issues/26869#issuecomment-2001925877
@@ -135,6 +59,26 @@ export default function Scanner() {
     return () => clearInterval(interval);
   }, []);
 
+  // Routes to next step if an image is selected
+  useEffect(() => {
+    if (selectedImage) {
+      router.push("/scan-receipt/parser");
+      router.setParams({ selectedImage });
+    }
+  }, [selectedImage]);
+
+  // Opens image library to select an image
+  const pickImageAsync = async () => {
+    const result = await ImagePicker.launchImageLibraryAsync({
+      allowsEditing: true,
+      quality: 1,
+    });
+
+    if (!result.canceled) {
+      setSelectedImage(result.assets[0].uri);
+    }
+  };
+
   if (hasPermission === null) {
     return <View />;
   }
@@ -145,10 +89,10 @@ export default function Scanner() {
   return (
     <ScrollView>
       <TouchableWithoutFeedback onPress={Keyboard.dismiss}>
-        <View style={{ flex: 1 }}>
+        <View style={styles.screen}>
           <Camera
             autoFocus={focus}
-            style={{ width: width, height: height * 0.8 }}
+            style={styles.camera}
             type={cameraType}
             ref={(ref) => {
               camera = ref;
@@ -164,25 +108,19 @@ export default function Scanner() {
                   )
                 }
               >
-                <Ionicons name="camera-reverse" size={30} color="white" />
+                <Ionicons name="camera-reverse" size={40} color="white" />
               </Pressable>
-              <Link
-                push
-                href={{
-                  pathname: "/scan-receipt/parser",
-                  params: { capturedImage, output },
+              <Pressable
+                onPress={() => {
+                  takePicture();
                 }}
-                asChild
+                style={styles.takePicButton}
               >
-                <Pressable
-                  onPress={() => {
-                    takePicture();
-                  }}
-                  style={styles.takePicButton}
-                >
-                  <Text style={styles.takePicText}>Take Picture</Text>
-                </Pressable>
-              </Link>
+                <Ionicons name="scan-circle" size={100} color="white" />
+              </Pressable>
+              <Pressable onPress={() => pickImageAsync()}>
+                <Ionicons name="images" size={35} color="white" />
+              </Pressable>
             </View>
           </Camera>
         </View>
@@ -192,23 +130,24 @@ export default function Scanner() {
 }
 
 const styles = StyleSheet.create({
-  takePicCont: {
-    flex: 1,
-    backgroundColor: "transparent",
+  screen: {},
+  camera: {
+    width: Dimensions.get("window").width,
+    height: Dimensions.get("window").height * 0.8,
     justifyContent: "flex-end",
+  },
+  takePicCont: {
+    flexDirection: "row",
+    backgroundColor: "transparent",
+    justifyContent: "center",
     alignItems: "center",
-    gap: 10,
+    marginBottom: 50,
+    gap: 40,
   },
   takePicButton: {
     borderWidth: 1,
     borderColor: "white",
-    borderRadius: 25,
-    marginBottom: 20,
-    paddingHorizontal: 20,
-    paddingVertical: 10,
-  },
-  takePicText: {
-    color: "white",
-    fontSize: 20,
+    borderRadius: 100,
+    padding: 5,
   },
 });
