+++ app/(tabs)/scan-receipt/parser.tsx
@@ -1,4 +1,4 @@
-import { Stack, useLocalSearchParams } from "expo-router";
+import { useLocalSearchParams } from "expo-router";
 import { useState } from "react";
 import {
   Button,
@@ -7,7 +7,6 @@ import {
   ScrollView,
   StyleSheet,
   Text,
-  TextInput,
   TouchableWithoutFeedback,
   View,
 } from "react-native";
@@ -15,7 +14,86 @@ import {
 export default function Parser() {
   const [text, setText] = useState("");
   const [recipe, setRecipe] = useState<string>("");
-  const { capturedImage, parsedOutput } = useLocalSearchParams();
+  const { selectedImage } = useLocalSearchParams();
+
+  const parsePicture = async () => {
+    if (selectedImage) {
+      const apiKey = "AIzaSyDFOfZ6SLPXEoDvF7RqdML5NXxOfySeKa4"; // Replace with your API key
+      const endpoint = `https://vision.googleapis.com/v1/images:annotate?key=${apiKey}`;
+      const base64Image = await convertImageToBase64(selectedImage[0]);
+
+      fetch(endpoint, {
+        method: "POST",
+        body: JSON.stringify({
+          requests: [
+            {
+              image: {
+                content: base64Image,
+              },
+              features: [
+                {
+                  type: "TEXT_DETECTION",
+                },
+              ],
+            },
+          ],
+        }),
+      })
+        .then((response) => response.json())
+        .then((data) => {
+          const annotations = data.responses[0].textAnnotations;
+          if (annotations && annotations.length > 0) {
+            // setText(annotations[0].description);
+            setText(parseHEBReceipt(annotations[0].description)); // Parse the detected text
+          } else {
+            // setText("No text detected");
+            setText("");
+          }
+        })
+        .catch((error) => console.error("Error:", error));
+    }
+  };
+
+  const convertImageToBase64 = async (imageUri: string) => {
+    const response = await fetch(imageUri);
+    const blob = await response.blob();
+    const base64String = await blobToBase64(blob);
+    return base64String;
+  };
+
+  const blobToBase64 = async (blob: Blob) => {
+    return new Promise<string>((resolve, reject) => {
+      const reader = new FileReader();
+      reader.onerror = reject;
+      reader.onload = () => {
+        const result: string | null = reader.result as string;
+        if (result) {
+          const [, base64Data] = result.split(",");
+          resolve(base64Data);
+        } else {
+          reject(new Error("Failed to read file as base64."));
+        }
+      };
+
+      reader.readAsDataURL(blob);
+    });
+  };
+
+  // Parses HEB receipt text
+  const parseHEBReceipt = (text: string) => {
+    // RegExp to match 1-2 digits followed by whitespace followed by any characters except "Ea."
+    const regex = /(^\d{1,2})\s+((?!Ea\.).*)/gm;
+    const matches = [...text.matchAll(regex)];
+
+    // Maps the matches to an array to print name
+    const items = matches
+      .map((item, index) => {
+        const [, , name] = item;
+        return `${index + 1}) ${name}`;
+      })
+      .join("\n");
+    return items;
+  };
 
   const generateRecipes = async () => {
     try {
@@ -52,16 +130,17 @@ export default function Parser() {
   return (
     <ScrollView>
       <TouchableWithoutFeedback onPress={Keyboard.dismiss}>
-        {capturedImage ? (
+        {selectedImage ? (
           <View style={{ alignItems: "center" }}>
             <Image
-              source={{ uri: capturedImage.toString() }}
+              source={{ uri: selectedImage.toString() }}
               style={{ width: 300, height: 300, marginBottom: 10 }}
             />
+            <Button title="Parse Image" onPress={parsePicture} />
             <Text style={{ marginTop: 20, fontSize: 18, fontWeight: "bold" }}>
               New Ingredients:
             </Text>
-            <Text>{parsedOutput}</Text>
+            <Text>{text}</Text>
             <Button title="Generate Recipes" onPress={generateRecipes} />
             <Text style={{ marginTop: 20, fontSize: 18, fontWeight: "bold" }}>
               Generated Recipe:
