+++ package.json
@@ -21,7 +21,8 @@
     "react-native": "0.73.6",
     "react-native-elements": "^3.4.3",
     "react-native-safe-area-context": "4.8.2",
-    "react-native-screens": "~3.29.0"
+    "react-native-screens": "~3.29.0",
+    "expo-image-picker": "~14.7.1"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
